Fix task 2 and 3 example python commands to work in both VM versions
The example ascii<-->hex python commands as originally shown in the lab description only work in the 16.04 VM. This proposed change adds a second command to task 2 that works in the 20.04 VM and changes the task 3 command to one that works in both VMs. A non-python alternative, which is consistent with the 3rd ed. of the books and also works in both VMs, is xxd. This proposed alternative is added in the form of a comment block in each task description.