Update Crypto_Padding_Oracle.tex
Fixed an issue in my previous commit where it mentioned that PKCS#5 works for block sizes less than 8 bytes - correct it to mention it only works for block sizes of 8 bytes exactly. Also added a new question to end of Task 1, asking students to explain why a full block of padding is added for the 16 byte file, and why this is necessary.